% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisquareTest.R
\name{chiSquareTest}
\alias{chiSquareTest}
\title{Chisquare test}
\usage{
chiSquareTest(model, type = c("dendrite", "age", "both"),
  ageDatasetCSV = NULL)
}
\arguments{
\item{model}{a Mclust object}

\item{type}{a string denoting dendrite compartment ("dendrite"), age ("age") or combination of both ("both")}

\item{ageDatasetCSV}{a path to a csv file where first column has the name of a dendrite and second column has the binary representation of the dendrite}
}
\value{
p-value of the test
}
\description{
Chisquare test to test independence between cluster distribution and dendritic compartment, age or combination of both
}
\examples{
chiSquareTest(model,type="age",system.file("extdata", "ageDataset.csv", package = "spineSimulation"))

}
