% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisquareTest.R
\name{chiSquareTestCluster}
\alias{chiSquareTestCluster}
\title{Chisquare test by cluster}
\usage{
chiSquareTestCluster(model, type = c("dendrite", "age", "both"),
  ageDatasetCSV = NULL)
}
\arguments{
\item{model}{a Mclust object}

\item{type}{a string denoting dendrite compartment ("dendrite"), age ("age") or combination of both ("both")}

\item{ageDatasetCSV}{a path to a csv file where first column has the name of a dendrite and second column has the binary representation of the dendrite}
}
\value{
p-value of the test
}
\description{
Chisquare test to test independence between cluster distribution and dendritic compartment, age or combination of both for each cluster
}
