% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance2soma.R
\name{chiSquareTestDistance}
\alias{chiSquareTestDistance}
\title{Chisquare distance to soma}
\usage{
chiSquareTestDistance(model, distanceCSV, numIntervals)
}
\arguments{
\item{model}{a Mclust object}

\item{distanceCSV}{path to a CSV file where distances to soma are saved}

\item{numIntervals}{is the number of intervals}
}
\value{
None
}
\description{
Chisquare test checking if the distribution of clusters is independent of the distance to soma
}
\examples{
chiSquareTestDistance(model,distanceCSV=system.file("extdata", "distanceSpines2soma.csv", package = "spineSimulation"),6)

}
