% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMDS.R
\name{computeMDS}
\alias{computeMDS}
\title{Compute multidimensional scaling}
\usage{
computeMDS(model, ndim)
}
\arguments{
\item{model}{is a Mclust object}

\item{ndim}{number of dimension t}
}
\value{
a list where first element is a matrix that represents the centroids of each cluster in a ndim-dimensional space and the second element is a matrix where each instance is placed in a ndim-dimensional space
}
\description{
Compute distance between clusters and place instances according to probability to belong to each cluster
}
\examples{
computeMDS(model,2)

}
