% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeOverlapping.R
\name{computeOverlapping}
\alias{computeOverlapping}
\title{Overlapping between pairs of clusters}
\usage{
computeOverlapping(model)
}
\arguments{
\item{model}{a Mclust object}
}
\value{
Table representing the overlapping between each pair of clusters
}
\description{
Measure the overlapping between each pair of multivariate Gaussians representing the clusters
}
\examples{
computeOverlapping(model)

}
