% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateArff.R
\name{generateArff}
\alias{generateArff}
\title{Write to Arff file}
\usage{
generateArff(dirArff, model)
}
\arguments{
\item{dirArff}{is the path to a folder where Arff files are saved}

\item{model}{is a Mclust object}
}
\value{
None
}
\description{
Write data of each cluster in a Arff file to read from Weka.
}
