% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotDistributions.R
\name{plotAge}
\alias{plotAge}
\title{Barplot spine distribution by age}
\usage{
plotAge(model, ageDatasetCSV)
}
\arguments{
\item{model}{a Mclust object}

\item{ageDatasetCSV}{a path to a csv file where first column has the name of a dendrite and second column has the binary representation of the dendrite}
}
\value{
None
}
\description{
Barplot of the distributions of the spines by their age and cluster
}
\examples{
plotAge(model,system.file("extdata", "ageDataset.csv", package = "spineSimulation"))

}
