% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotDistributions.R
\name{plotCombination}
\alias{plotCombination}
\title{Barplot spine distribution by combination of dendritic compartment and age}
\usage{
plotCombination(model, ageDatasetCSV)
}
\arguments{
\item{model}{a Mclust object}

\item{ageDatasetCSV}{a path to the csv where is saved the age of each dendrite}
}
\value{
None
}
\description{
Barplot of the distributions of the spines by combination of dendritic compartment and age
}
\examples{
plotCombination(model,system.file("extdata", "ageDataset.csv", package = "spineSimulation"))

}
