% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotDistributions.R
\name{plotDendriticCompartment}
\alias{plotDendriticCompartment}
\title{Barplot spine distribution by dendritic compartment}
\usage{
plotDendriticCompartment(model)
}
\arguments{
\item{model}{a Mclust object}
}
\value{
None
}
\description{
Barplot of the distributions of the spines by their dendritic compartment and cluster
}
\examples{
plotDendriticCompartment(model)

}
