% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance2soma.R
\name{plotDistance2Soma}
\alias{plotDistance2Soma}
\title{Barplot distance to soma}
\usage{
plotDistance2Soma(model, distanceCSV, numIntervals)
}
\arguments{
\item{model}{a Mclust object}

\item{distanceCSV}{path to a CSV file where distances to soma are saved}

\item{numIntervals}{is the number of intervals}
}
\value{
None
}
\description{
Barplot representing the distribution of the cluster according to their distance to soma divided in intervals
}
\examples{
plotDistance2Soma(model,distanceCSV=system.file("extdata", "distanceSpines2soma.csv", package = "spineSimulation"),6)

}
