% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotDistributions.R
\name{plotGlobalDistribution}
\alias{plotGlobalDistribution}
\title{Barplot spine distribution by cluster}
\usage{
plotGlobalDistribution(model)
}
\arguments{
\item{model}{a Mclust object}
}
\value{
None
}
\description{
Barplot of the distributions of the spines by cluster according to a given model
}
\examples{
plotGlobalDistribution(model)

}
