% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMDS.R
\name{plotMDS}
\alias{plotMDS}
\title{Plot result of multidimensional scaling}
\usage{
plotMDS(model, MDS, colors = NULL)
}
\arguments{
\item{model}{an Mclust object}

\item{MDS}{a list of matrices obtained from computeMDS}

\item{colors}{a matrix of rgb colors where each row defines the colors of a cluster}
}
\value{
None
}
\description{
Plot the 2D or 3D multidimensional scaling result.
}
\examples{
MDS<-computeMDS(model,2)
plotMDS(model,MDS)

}
