% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalVariance.R
\name{plotTotalVariance}
\alias{plotTotalVariance}
\title{Barplot total variance (log)}
\usage{
plotTotalVariance(model)
}
\arguments{
\item{model}{a Mclust object}
}
\value{
None
}
\description{
Barplot representing the log of the total variance of each cluster
}
\examples{
plotTotalVariance(model)

}
