% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSpines.R
\name{spineSampling}
\alias{spineSampling}
\title{Simulate new spines}
\usage{
spineSampling(model, nSpines = 100, cluster = 0, seed = 1)
}
\arguments{
\item{model}{is a Mclust object}

\item{nSpines}{is an integer that denotes the number of spines that must be simulated}

\item{cluster}{is an integer value that denotes the cluster from it should be sampled new spines, when it is 0 means that it should be sampled from prior}

\item{seed}{is an integer value used to obtain reproducible results}
}
\value{
a dataframe with then nSpines simulated spines
}
\description{
Simulate virtual spines from the mixture of multivariate Gaussians
}
\examples{
newSpines <- spineSampling(model,nSpines=100,cluster=0,seed=1)

}
