\name{spineSimulation-package}

\alias{spineSimulation}

\title{3D morphological clustering and simulation of human dendritic spines}
\description{
Morphology of dendritic spines appears to be critical from the functional point of view. Although many different classifications of spines have been proposed, the most widely used categorization is still that of Peters and Kaiserman-Abramof, which traditionally groups spines into four types (thin, mushroom, stubby and filopodium). In the present work we used over 7000 individually 3D reconstructed dendritic spines from human cortical pyramidal neurons to cluster spine morphologies. This approach uncovered nine distinct groups of human spines. To better understanding the differences among them, the discriminant characteristics of each one were identified as a set of rules. The model-based clustering also allowed to simulate realistic 3D virtual representations of spines that matched morphological definitions of each cluster.

We introduce a method to simulate three-dimensional spines from a model-based clustering with multivariate Gaussian distributions. Clustering results are interpreted with tools like multidimensional scaling based on Hellinger distance, probability distributions and overlapping between clusters.
}

\details{
Package: spineSimulation

Type: Package

Version: 1.0

License: GPL(>=2)
}

\author{
Sergio Luengo-Sanchez, et al.
}

\references{
[1] 3D morphological clustering and simulation of human dendritic spines
}
