% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdBelonging.R
\name{thresholdBelonging}
\alias{thresholdBelonging}
\title{Get the numbers of spines with a belonging probability under a threshold}
\usage{
thresholdBelonging(model)
}
\arguments{
\item{a}{Mclust object}
}
\value{
num_spines_over_threshold a table with a probability of belonging lower than a threshold
}
\description{
Given a threshold probability count the number of spines under that bound
}
\examples{
thresholdBelonging(model)

}
